refactor: User의 컨트롤러와 서비스를 interface 기반으로 분리
